refactor: move API-related classes to package 'api'
