+++ frontend/app/src/main/java/com/example/gathernow/api/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.example.gathernow;
+package com.example.gathernow.api;
 
 public class TokenResponse {
     private String message;
