+++ frontend/app/src/test/java/com/example/gathernow/LogInActivityTest.java
@@ -10,9 +10,9 @@ import retrofit2.Call;
 
 import static org.mockito.Mockito.when;
 
-import com.example.gathernow.CodeMessageResponse;
+import com.example.gathernow.api.CodeMessageResponse;
 import com.example.gathernow.LogInActivity;
-import com.example.gathernow.ServiceApi;
+import com.example.gathernow.api.ServiceApi;
 import com.example.gathernow.UserData;
 
 @RunWith(MockitoJUnitRunner.class)
